requirements.txt
streamlit==1.37.1
pandas==2.2.2
plotly==5.17.0
gspread==6.2.1
google-auth==2.40.3